+++ test/next/jdbc/test_fixtures.clj
@@ -4,6 +4,7 @@
   "Multi-database testing fixtures."
   (:require [clojure.string :as str]
             [next.jdbc :as jdbc]
+            [next.jdbc.prepare :as prep]
             [next.jdbc.sql :as sql])
   (:import (com.opentable.db.postgres.embedded EmbeddedPostgres)
            (javax.sql DataSource)))