+++ CHANGELOG.md
@@ -6,7 +6,9 @@ Only accretive/fixative changes will be made from now on.
 
 The following changes have been committed to the **master** branch since the 1.0.12 release:
 
-* None.
+* Fix #82 by adding `clojure.java.data`-based support for setting arbitrary properties on `Connection` and `PreparedStatement` objects, post-creation. Note: this uses the Java reflection API under the hood.
+* Adds `next.jdbc.prepare/statement` to create `Statement` objects with all the options available to `prepare` except `:return-keys`.
+* Update `org.clojure/java.data` to 0.1.5 (for property setting).
 
 ## Stable Builds